refactor: extract helpers and auth from server.go
